fix: move parallel event to types
- Divide types by modules
