+++ blocks/types/__init__.py
@@ -1,0 +1,10 @@
+from blocks.types.base import (Event,
+                               EventOrEvents,
+                               Processor,
+                               AsyncProcessor,
+                               AnyProcessor,
+                               Source,
+                               AsyncSource,
+                               AnySource,
+                               Block)
+from blocks.types.dto import ParallelEvent
